A star

+++ entities/JsonHandlers/Serializers/CoordinatesJsonSerializer.cpp
@@ -5,10 +5,10 @@ CoordinatesJsonSerializer::toJson(Coordinates coordinates, rapidjson::Document::
     rapidjson::Value json;
     json.SetArray();
     rapidjson::Value lonVal;
-    lonVal.SetDouble(coordinates.longitude());
+    lonVal.SetDouble((double)coordinates.longitude());
     json.PushBack(lonVal, allocator);
     rapidjson::Value latVal;
-    latVal.SetDouble(coordinates.latitude());
+    latVal.SetDouble((double)coordinates.latitude());
     json.PushBack(latVal, allocator);
     return json;
 }
